Fix some type casting warnings.

+++ apps/freeablo/falevelgen/levelgen.cpp
@@ -42,7 +42,7 @@ namespace FALevelGen
                 {
                     if((xCoord == x + xPos && yCoord == yPos) ||
                        (xCoord == x + xPos && yCoord == height-1 + yPos))
-                        return true; 
+                        return true;
                     //level[x + room.xPos][room.yPos] = wall;
                     //level[x + room.xPos][room.height-1 + room.yPos] = wall;
                 }
@@ -52,7 +52,7 @@ namespace FALevelGen
                 {
                     if((xCoord == xPos && yCoord == y + yPos) ||
                        (xCoord == width-1 + xPos && yCoord == y + yPos))
-                        return true; 
+                        return true;
 
 
                     //level[room.xPos][y + room.yPos] = wall;
@@ -74,7 +74,7 @@ namespace FALevelGen
 
             size_t distance(const Room& other) const
             {
-                return sqrt((float)((centre().first - other.centre().first)*(centre().first - other.centre().first) + (centre().second - other.centre().second)*(centre().second - other.centre().second)));
+                return static_cast<size_t> (sqrt(static_cast<float>((centre().first - other.centre().first) * (centre().first - other.centre().first) + (centre().second - other.centre().second) * (centre().second - other.centre().second))));
             }
     };
 
@@ -112,7 +112,7 @@ namespace FALevelGen
                 fillRoom(corridoorRooms[i], level);
         }
     }
-    
+
     // Ensures that two points are connected by inserting an l-shaped corridoor,
     // also draws any rooms in rooms vector that the corridoor intersects
     void connect(const Room& a, const Room& b, const std::vector<Room>& rooms, Level::Dun& level)
@@ -122,7 +122,7 @@ namespace FALevelGen
 
         size_t bx = b.centre().first;
         size_t by = b.centre().second;
-        
+
         if(bx != ax)
         {
             if(bx > ax)
@@ -132,7 +132,7 @@ namespace FALevelGen
             }
             else
             {
-                Room x(bx, ay-1, ax-bx, 3); 
+                Room x(bx, ay-1, ax-bx, 3);
                 drawCorridoorSegment(x, rooms, level);
             }
         }
@@ -147,12 +147,12 @@ namespace FALevelGen
             {
                 Room y(bx-1, by, 3, ay-by+2);
                 drawCorridoorSegment(y, rooms, level);
-                
+
             }
 
         }
     }
-   
+
     // Move room in direction specified by normalised vector, making sure to keep within
     // grid of size width * height
     void moveRoom(Room& room, const std::pair<float, float>& vector, size_t width, size_t height)
@@ -220,7 +220,7 @@ namespace FALevelGen
                 assert(false); // This should never happen
             }
         }
-        
+
         int32_t newX = room.xPos + xMove;
         int32_t newY = room.yPos + yMove;
 
@@ -238,7 +238,7 @@ namespace FALevelGen
         vector.first /= (float)magnitude;
         vector.second /= (float)magnitude;
     }
-    
+
     // Removes the room overlapping the largest number of rooms repeatedly,
     // until there are no overlaps
     void removeOverlaps(std::vector<Room>& rooms)
@@ -264,7 +264,7 @@ namespace FALevelGen
                         overlap = true;
                     }
                 }
-                
+
                 if(neighbourCount > maxNeighbourCount)
                 {
                     maxIndex = i;
@@ -276,9 +276,9 @@ namespace FALevelGen
                 rooms.erase(rooms.begin() + maxIndex);
         }
     }
-    
+
     // Separate rooms so they don't overlap, using flocking ai
-    // based on the algorithm described here: 
+    // based on the algorithm described here:
     // http://gamedevelopment.tutsplus.com/tutorials/the-three-simple-rules-of-flocking-behaviors-alignment-cohesion-and-separation--gamedev-3444
     void separate(std::vector<Room>& rooms, size_t width, size_t height)
     {
@@ -294,17 +294,17 @@ namespace FALevelGen
 
             for(size_t i = 0; i < rooms.size(); i++)
             {
-                std::pair<float, float> vector(0, 0);
+                std::pair<float, float> vector(0.f, 0.f);
 
                 std::pair<int32_t, int32_t> currentCentre = rooms[i].centre();
-                
+
                 size_t neighbourCount = 0;
 
                 for(size_t j = 0; j < rooms.size(); j++)
                 {
                     if(i == j)
                         continue;
-                    
+
                     bool intersects = rooms[i].intersects(rooms[j]);
 
                     overlap = overlap || intersects;
@@ -315,12 +315,12 @@ namespace FALevelGen
 
                         if(centre.first == currentCentre.first && centre.second == currentCentre.second)
                         {
-                            vector.first = randomInRange(0, 10);
-                            vector.second = randomInRange(0, 10);
+                            vector.first = static_cast<float> (randomInRange(0, 10));
+                            vector.second = static_cast<float> (randomInRange(0, 10));
                             neighbourCount++;
                             continue;
                         }
-                        
+
                         std::pair<float, float> iToJ = Misc::getVec(currentCentre, centre);
                         normalise(iToJ);
 
@@ -333,23 +333,23 @@ namespace FALevelGen
 
                 if(vector.first == 0 && vector.second == 0)
                     continue;
-               
+
                 vector.first /= (float)neighbourCount;
                 vector.second /= (float)neighbourCount;
 
                 normalise(vector);
-                
+
                 // invert
                 vector.first *= -1;
                 vector.second *= -1;
-                
+
                 moveRoom(rooms[i], vector, width, height);
             }
 
         }
-        
+
         if(overlap)
-            removeOverlaps(rooms);    
+            removeOverlaps(rooms);
     }
 
     void generateRooms(std::vector<Room>& rooms, size_t width, size_t height)
@@ -360,11 +360,11 @@ namespace FALevelGen
 
         size_t centreX = width/2;
         size_t centreY = height/2;
-       
+
         // The following two are based on the fact that 150 rooms in a radius of
-        // 15 looks good on an 85*75 map 
-        size_t numRooms = std::min(width, height)*(150.0/75.0);
-        size_t radius = std::min(width, height)*(15.0/75.0);
+        // 15 looks good on an 85*75 map
+        size_t numRooms = static_cast<size_t> (std::min(width, height)*(150.0/75.0));
+        size_t radius = static_cast<size_t> (std::min(width, height)*(15.0/75.0));
 
         while(placed < numRooms)
         {
@@ -372,10 +372,10 @@ namespace FALevelGen
 
             if(((centreX-newRoom.centre().first)*(centreX-newRoom.centre().first) + (centreY-newRoom.centre().second)*(centreY-newRoom.centre().second)) > radius*radius)
                 continue;
-            
+
             newRoom.width = normRand(4, std::min(width-newRoom.xPos, maxDimension));
             newRoom.height = normRand(4, std::min(height-newRoom.yPos, maxDimension));
-            
+
             float ratio = ((float)newRoom.width) / ((float)newRoom.height);
 
             if(ratio < 0.5 || ratio > 2.0)
@@ -403,7 +403,7 @@ namespace FALevelGen
             level[room.xPos][y + room.yPos] = wall;
             level[room.width-1 + room.xPos][y + room.yPos] = wall;
         }
-        
+
         // Fill ground
         for(size_t x = 1; x < room.width-1; x++)
         {
@@ -413,16 +413,16 @@ namespace FALevelGen
             }
         }
     }
-    
+
     // Get the value at (x,y) in level, or zero if it is an invalid position
     size_t getXY(int32_t x, int32_t y, const Level::Dun& level)
     {
         if(x < 0 || x >= (int32_t)level.width() || y < 0 || y >= (int32_t)level.height())
             return 0;
-        
+
         return level[x][y];
     }
-    
+
     // Returns true if the tile at (x,y) in level borders any tile of the value tile,
     // false otherwise
     bool borders(size_t x, size_t y, Basic tile, const Level::Dun& level)
@@ -433,12 +433,12 @@ namespace FALevelGen
             {
                 int32_t testX = xoffs + x;
                 int32_t testY = yoffs + y;
-                
+
                 if(getXY(testX, testY, level) == tile)
                     return true;
             }
         }
-        
+
         return false;
     }
 
@@ -560,7 +560,7 @@ namespace FALevelGen
             }
         }
     }
-    
+
     // Remove double walls, as the tileset does not allow for them
     void cleanup(Level::Dun& level, std::vector<Room>& rooms)
     {
@@ -577,7 +577,7 @@ namespace FALevelGen
 
         cleanLooseWalls(level, false);
     }
-    
+
     // Helper function for adding doors
     // Iterates over all blocks on a wall, and adds doors where necessary, looking at what is in the direction
     // indicated by add (1 or -1) to determine if a door is needed
@@ -589,13 +589,13 @@ namespace FALevelGen
 
         for(size_t i = start; i < end; i++)
         {
-            if((xAxis && (getXY(i, otherCoord, level) == floor || getXY(i, otherCoord, level) == door)) || 
+            if((xAxis && (getXY(i, otherCoord, level) == floor || getXY(i, otherCoord, level) == door)) ||
               (!xAxis && (getXY(otherCoord, i, level) == floor || getXY(otherCoord, i, level) == door)))
             {
                 hole = true;
             }
             else if((xAxis && getXY(i, otherCoord+add, level) != floor) ||
-                   (!xAxis && getXY(otherCoord+add, i, level) != floor)) 
+                   (!xAxis && getXY(otherCoord+add, i, level) != floor))
             {
                 if(hole)
                     region.resize(0);
@@ -611,12 +611,12 @@ namespace FALevelGen
                     {
                         if(region.size() > 0)
                             level[region[region.size()/2].first][region[region.size()/2].second] = door;
-                        
+
                         region.resize(0);
                         connected = true;
                     }
-                    
-                    if(xAxis) 
+
+                    if(xAxis)
                     {
                         if(getXY(i-1, otherCoord, level) == wall && getXY(i+1, otherCoord, level) == wall)
                             region.push_back(std::pair<size_t, size_t>(i, otherCoord));
@@ -633,24 +633,24 @@ namespace FALevelGen
                 }
             }
         }
-        
+
         if(!hole && region.size() > 0)
             level[region[region.size()/2].first][region[region.size()/2].second] = (levelNum == 4) ? floor : door;
     }
-    
+
     void addDoors(Level::Dun& level, const std::vector<Room>& rooms, size_t levelNum)
     {
         for(size_t i = 0; i < rooms.size(); i++)
         {
             // Top x wall
-            doorAddHelper(level, rooms[i].yPos,                   -1, rooms[i].xPos+1, rooms[i].xPos + rooms[i].width -1,  true, levelNum); 
+            doorAddHelper(level, rooms[i].yPos,                   -1, rooms[i].xPos+1, rooms[i].xPos + rooms[i].width -1,  true, levelNum);
             // Bottom x wall
-            doorAddHelper(level, rooms[i].yPos+rooms[i].height-1, +1, rooms[i].xPos+1, rooms[i].xPos + rooms[i].width -1,  true, levelNum); 
-            
+            doorAddHelper(level, rooms[i].yPos+rooms[i].height-1, +1, rooms[i].xPos+1, rooms[i].xPos + rooms[i].width -1,  true, levelNum);
+
             // Left y wall
-            doorAddHelper(level, rooms[i].xPos,                   -1, rooms[i].yPos+1, rooms[i].yPos + rooms[i].height -1, false, levelNum); 
+            doorAddHelper(level, rooms[i].xPos,                   -1, rooms[i].yPos+1, rooms[i].yPos + rooms[i].height -1, false, levelNum);
             // Right y wall
-            doorAddHelper(level, rooms[i].xPos+rooms[i].width-1,  +1, rooms[i].yPos+1, rooms[i].yPos + rooms[i].height -1, false, levelNum); 
+            doorAddHelper(level, rooms[i].xPos+rooms[i].width-1,  +1, rooms[i].yPos+1, rooms[i].yPos + rooms[i].height -1, false, levelNum);
         }
     }
 
@@ -662,8 +662,8 @@ namespace FALevelGen
             {
                 size_t baseX = rooms[i].xPos + (rooms[i].width/2);
                 size_t baseY = rooms[i].yPos;
-                
-                // on a wall       
+
+                // on a wall
                 if(level[baseX-1][baseY-2] == blank && level[baseX][baseY-2] == blank && level[baseX+1][baseY-2] == blank &&
                    level[baseX-1][baseY-1] == blank && level[baseX][baseY-1] == blank && level[baseX+1][baseY-1] == blank &&
                    level[baseX-1][baseY] == wall && level[baseX][baseY] == wall && level[baseX+1][baseY] == wall &&
@@ -687,12 +687,12 @@ namespace FALevelGen
                         continue;
 
                     level[baseX][baseY] = upStairs;
-                    
+
                     return true;
                 }
             }
         }
-            
+
 
         return false;
     }
@@ -728,7 +728,7 @@ namespace FALevelGen
                         continue;
 
                 level[baseX][baseY] = downStairs;
-                
+
                 return true;
             }
         }
@@ -737,24 +737,24 @@ namespace FALevelGen
     }
 
     #define ROOMAREA 30
-    
+
     // Generates a flat map (no information about wall direction, etc)
     // Uses the tinykeep level generation algorithm, described here:
     // http://www.reddit.com/r/gamedev/comments/1dlwc4/procedural_dungeon_generation_algorithm_explained/
     // The basic algorithm is as follows:
-    //     1. Generate a bunch of rooms in a radius around the centre of the map, with rooms weighted 
+    //     1. Generate a bunch of rooms in a radius around the centre of the map, with rooms weighted
     //        towards being small more often than large.
     //     2. Use separation steering to spread them out until they no longer overlap.
     //     3. Split the rooms into two types, real rooms, and corridoor rooms, where real rooms are rooms
     //        with an area above a certain threshold, and corridoor rooms are the rest.
     //     4. Construct a minimum spanning tree which connects all the rooms together, then add in some
     //        extra edges to allow for some loops.
-    //     5. Connect the rooms according to the graph from the last step with l shaped corridoors, and 
+    //     5. Connect the rooms according to the graph from the last step with l shaped corridoors, and
     //        also draw any corridoor rooms that the corridoors overlap as part of the corridoor.
     Level::Dun generateTmp(size_t width, size_t height, size_t levelNum)
     {
         Level::Dun level(width, height);
-        
+
         // Initialise whole dungeon to blank
         for(size_t x = 0; x < width; x++)
             for(size_t y = 0; y < height; y++)
@@ -763,7 +763,7 @@ namespace FALevelGen
         std::vector<Room> rooms;
         std::vector<Room> corridoorRooms;
         generateRooms(rooms, width, height);
-        
+
         // Split rooms into real rooms, and corridoor rooms
         for(size_t i = 0; i < rooms.size(); i++)
         {
@@ -774,7 +774,7 @@ namespace FALevelGen
                 i--;
             }
         }
-        
+
         // Create graph with edge from each room to each other room
         std::vector<std::vector<size_t> > graph(rooms.size());
         for(size_t i = 0; i < rooms.size(); i++)
@@ -790,33 +790,33 @@ namespace FALevelGen
         minimumSpanningTree(graph, parent);
         for(size_t i = 1; i < rooms.size(); i++)
             connect(rooms[parent[i]], rooms[i], corridoorRooms, level);
-        
+
         // Add in an extra 15% of the number of rooms random connections to create some loops
-        size_t fifteenPercent = (((float)rooms.size())/100.0)*15.0;
+        size_t fifteenPercent = static_cast<size_t> (((rooms.size())/100.0)*15.0);
         for(size_t i = 0; i < fifteenPercent; i++)
         {
             size_t a, b;
-            
+
             do
             {
                 a = randomInRange(0, rooms.size()-1);
                 b = randomInRange(0, rooms.size()-1);
             }
             while(a == b || parent[a] == b || parent[b] == a);
-            
+
             connect(rooms[a], rooms[b], corridoorRooms, level);
         }
 
         // Draw rooms on top of corridoors
         for(size_t i = 0; i < rooms.size(); i++)
             drawRoom(rooms[i], level);
-        
+
         // Bound corridoors with walls
         addWalls(level);
 
-        cleanup(level, rooms); 
+        cleanup(level, rooms);
         addDoors(level, rooms, levelNum);
-        
+
         // Make sure we always place stairs
         if(!(placeUpStairs(level, rooms, levelNum) && placeDownStairs(level, rooms, levelNum)))
             return generateTmp(width, height, levelNum);
@@ -840,13 +840,13 @@ namespace FALevelGen
     {
         return getXY(x, y, tmpLevel) == floor || getXY(x, y, tmpLevel) == door;
     }
- 
+
     void setPoint(int32_t x, int32_t y, size_t val, const Level::Dun& tmpLevel,  Level::Dun& level, int32_t wallOffset, bool isInsideWall)
     {
         size_t newVal = val;
 
         if(val == (size_t)TileSetEnum::xWall+wallOffset)
-        {   
+        {
             if(getXY(x, y, tmpLevel) == door && isInsideWall)
                 newVal = TileSetEnum::xDoor;
             else if(getXY(x, y, tmpLevel) == upStairs)
@@ -864,7 +864,7 @@ namespace FALevelGen
         }
 
         else if(val == (size_t)TileSetEnum::yWall+wallOffset)
-        {   
+        {
             if(getXY(x, y, tmpLevel) == door && isInsideWall)
                 newVal = TileSetEnum::yDoor;
             else if(!isInsideWall && getXY(x+1, y, tmpLevel) == blank)
@@ -908,11 +908,11 @@ namespace FALevelGen
 
         level[x][y] = newVal;
     }
-    
+
     void placeMonsters(Level::Level& level, std::vector<FAWorld::Actor*>& actors, const DiabloExe::DiabloExe& exe, size_t dLvl)
     {
         std::vector<const DiabloExe::Monster*> possibleMonsters = exe.getMonstersInLevel(dLvl);
-        
+
         for(size_t i = 0; i < (level.height() + level.width())/2; i++)
         {
             size_t xPos, yPos;
@@ -1114,7 +1114,7 @@ namespace FALevelGen
         }
     }
 
- 
+
     FAWorld::GameLevel* generate(size_t width, size_t height, size_t dLvl, const DiabloExe::DiabloExe& exe, size_t previous, size_t next)
     {
         size_t levelNum = ((dLvl-1) / 4) + 1;
@@ -1152,7 +1152,7 @@ namespace FALevelGen
                     level[x][y] = tileset.convert(((TileSetEnum::TileSetEnum)level[x][y]));
             }
         }
-        
+
         // Add in some random aesthetic variation
         for(int32_t x = 0; x < (int32_t)width; x++)
         {
@@ -1198,7 +1198,7 @@ namespace FALevelGen
             level[x][y+1] = tileset.downStairs8;
             level[x+1][y+1] = tileset.downStairs9;
         }
-        
+
         ss.str(""); ss << "levels/l" << levelNum << "data/l" << levelNum << ".cel";
         std::string celPath = ss.str();
 
@@ -1210,7 +1210,7 @@ namespace FALevelGen
 
         ss.str(""); ss << "levels/l" << levelNum << "data/l" << levelNum << ".sol";
         std::string solPath = ss.str();
-                
+
         Level::Level retval(level, tilPath, minPath, solPath, celPath, downStairsPoint, upStairsPoint, tileset.getDoorMap(), previous, next);
 
         std::vector<FAWorld::Actor*> actors;
